fix(formula-engine): guard coord i32 overflow + cap MAX_ROW
- Use i64 intermediates + explicit range checks in Coord::{normalize,to_a1_index,fmt_r1c1} to avoid wrapping when indices/origins exceed i32::MAX.
- When a relative offset cannot be represented in i32, keep the original A1 coord (no normalization) and fall back to absolute R1C1 rendering.
- Cap eval/compiler MAX_ROW to i32::MAX-1 to match eval::ast::Ref encoding; add regression tests incl. deterministic #REF! for out-of-range rows.

+++ crates/formula-engine/src/eval/compiler.rs
@@ -15,12 +15,12 @@ const MAX_COL: u32 = EXCEL_MAX_COLS - 1;
 
 /// Maximum row index supported by the engine (0-indexed).
 ///
-/// Rows are **not** capped here: references beyond Excel's default row count remain syntactically
-/// valid and are validated dynamically at evaluation time using per-sheet dimensions.
+/// Rows are capped by what the eval IR can encode: [`eval::ast::Ref`] stores absolute row/column
+/// components in an `i32` with [`eval::ast::Ref::SHEET_END`] (`i32::MAX`) reserved as a sentinel.
+/// As a result, the largest concrete row index we can compile is `i32::MAX - 1` (0-indexed).
 ///
-/// `u32::MAX` is reserved for the sheet-end sentinel used in whole-row / whole-column references,
-/// so the largest representable concrete row index is `u32::MAX - 1`.
-const MAX_ROW: u32 = u32::MAX - 1;
+/// References beyond this limit compile deterministically to `#REF!`.
+const MAX_ROW: u32 = i32::MAX as u32 - 1;
 
 fn parse_number(raw: &str) -> Option<f64> {
     match raw.parse::<f64>() {
